refactor: extract shared snapshot module from detector duplicates
